Added an error (on purpose) Fixed some code

+++ src/domain/model/Game.java
@@ -4,10 +4,10 @@ import domain.DomainException;
 
 public class Game {
     private int minimumLeeftijd;
-    private double prijs;
+    private int prijs;
     private final String naam;
 
-    public Game(String naam, int minimumLeeftijd, double prijs) {
+    public Game(String naam, int minimumLeeftijd, int prijs) {
         setMinimumLeeftijd(minimumLeeftijd);
         setPrijs(prijs);
         if (naam == null || naam.trim().isEmpty() || naam.trim().length() < 3)
@@ -29,7 +29,7 @@ public class Game {
         return prijs;
     }
 
-    public void setPrijs(double prijs) {
+    public void setPrijs(int prijs) {
         if (prijs<0)
             throw new DomainException("Geen geldige prijs");
         this.prijs = prijs;
